google_opinion_baseline/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
docker-pycreds==0.4.0
easydict==1.10
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.30
huggingface-hub==0.12.0
idna==3.4
joblib==1.2.0
lightning-utilities==0.6.0.post0
multidict==6.0.4
numpy==1.24.2
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
protobuf==3.20.1
psutil==5.9.4
python-dateutil==2.8.2
pytorch-lightning==1.9.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
sentencepiece==0.1.97
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tokenizers==0.13.2
# torch==1.13.1
torchmetrics==0.11.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.4.0
urllib3==1.26.14
wandb==0.13.10
yarl==1.8.2
